+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 Below are the noteworthy changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
-##  Upcoming release: 0.13.0a2 (2024-Oct-??)
+##  Upcoming release: 0.13.0a2 (2024-Oct-11)
 ### Migration of checks
 #### Moved from OpenType to Universal profile
   - **[name/no_copyright_on_description]**: Adding copyright information to the description name entry is bad practice but is not a breach of specification. (issue #4857)
   - **[name/italic_names]**: This check uses the filename to determine whether the font should have various italic bits set. File naming is a matter for the OpenType specification. (issue #4858)
 
 ### Changes to existing checks
+#### On the Universal Profile
+  - **[caps_vertically_centered]:** Reenabled and corrected it. If possible, the uppercase glyphs should be vertically centered in the em box. (issues #4139 and #4274)
+
 #### On the Google Fonts profile
   - **[googlefonts/article/images]:** Missing image files are now reported at **FATAL** level. (issue #4848)